core/requirements.txt
Django==5.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
Pillow==10.1.0
channels==4.0.0
asgiref==3.8.1
daphne==4.0.0
django-cors-headers==4.4.0
psycopg[binary]==3.2.3
